TLDR: Commented out console statement but kept for future testing.
There is a statement that prints to console the boolean value of `delete-on-leave`. This
statement can be used in testing to display the associated value, and confirm the logic
flow through the program. This has been commented as it may be useful for testing in the
future, when the team decides to use a config variable (or not) controlling the delete
function.