Uncommented previous tests and added DELETE_ON_LEAVE switch.
When implementing the feature--delete-on-leave, the lines that refreshed the browser
were commented to prevent errors from testing. The errors were due to failed assertions
of objects that no longer existed on refresh, because refreshing would cause board
deletion.

As part of the feature--delete-on-leave, a switch was added (`DELETE_ON_LEAVE`) that
when set to `false` would prevent board deletion. This has been enabled in this file
and the previously commented lines uncommented, returning the previous assertions to
this file and the original testing scheme.