Updated nightwatch.conf.js and added new delete-on-leave test suite
Due to Response 500 errors when running multiple test suites, the nightwatch
configuration file was updated to included parallelized workers. Due to running
multiple workers (i.e. threads) and therefore sessions, this resolves the 500
error as there is a dedicated server instance for each test suite (i.e. separate
file in `tests` directory).

Using the original `integration.js` test suite, a new `delete-on-leave.js` test
suite was added to the `tests` directory. This includes a simple test to confirm
that existing whiteboard elements are removed from the board on refresh.