Board deletion does not occur for two users test
Once the board is open in one window, another window is opened. The
element is checked to make sure it still exists, even after the window
is refreshed. The second window is closed, and elements confirmed to
still exist in the first window. The first window is refreshed, ending
the board session, and then the element confirmed as removed.

This test confirms that the board is not deleted/wiped until there are
all users have left.